Decompose posts
Також виправив елементи які не потребують закриття і добавив інлайн функції
