+++ src/react/posts.js
@@ -1,37 +1,21 @@
 import axios from 'axios';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useSelector } from 'react-redux';
 import crossicon from '../images/Crossicon.png';
-import avatar from "../images/avatar.jpg";
+import Post from "../components/post.js";
 
 function Posts() {
-  const [posts, setPosts] = useState([])
-  const [error, setError] =useState(null)
   const userId = useSelector((state) => state.auth.userId)
   const [textarea, setTextarea] = useState('')
   let [isOpen, setIsOpen] = useState(false) 
+  const redirect = () => setIsOpen(prevState => !prevState)
 
-  function redirect(){
-    setIsOpen(prevState => !prevState)
-  }
-
-  function submitHandler (e) {
+  const submitHandler = () => {
     axios.post('http://localhost:8080/createpost', {text: textarea, userId: userId})
-    .then((data) =>{
+    .then(() =>{
       setTextarea('')
     })
   }
-
-  useEffect(() =>{  
-    axios.get('http://localhost:8080/getposts')
-    .then(res =>{
-      let publications = []
-      res.data.forEach(post => publications.push(post))
-      setPosts(publications)
-      setError(null)  
-    })
-    .catch(err => setError('Couldn`t display posts'))
-  })
 
   return (
     <div className='flex min-h-screen bg-slate-100 justify-center pt-11'>
@@ -41,7 +25,7 @@ function Posts() {
             <div className='fixed bg-white w-1/3 left-1/3 top-1/4 rounded border shadow text-lg pb-4'>
               <div className='flex justify-between items-center border-b pl-1'>
                 <b>Create post</b>
-                <img className='size-10 m-1 rounded-full' onClick={redirect} src={crossicon} alt='Cancel'></img>
+                <img className='size-10 m-1 rounded-full' onClick={redirect} src={crossicon} alt='Cancel'/>
               </div>
               <form className="flex flex-col bg-white h-4/5">
                 <textarea className='text-2xl resize-none overflow-auto outline-none' rows={6} placeholder="what's on your mind?"  value={textarea} onChange={(e) => setTextarea(e.target.value)} required />
@@ -49,22 +33,10 @@ function Posts() {
               </form>
             </div>
         )}
-        {posts.map((post) =>{
-          
-          return <div className='bg-white mt-2 rounded shadow' key={post.postId}>
-            <div className='flex border-b'>
-              <img className='size-10 m-1 rounded-full' src={avatar} alt='avatar22'></img>
-              <div>
-                <p className='text-xl'>{post.FirstName + " " + post.LastName}</p>
-                <p className='text-sm text-gray-400'>{post.postDate}</p>
-              </div>
-            </div>
-            <p className='p-2 text-2xl break-words'>{post.postValue}</p>
-            </div>
-        })}
-          {error ? <p>{error}</p>: null}
+        <Post/>
       </div>
     </div>
-)}
+  )
+}
 
 export default Posts;
